moved types to shared folder
