+++ frontend/shop-floor-assistance-frontend/src/app/components/editor-order/editor-order.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { OrderTableComponent } from '../../shared/component-elements/order-table/order-table.component';
 import { ButtonComponent } from '../../shared/component-elements/button/button.component';
 import { Router, RouterLink } from '@angular/router';
-import { orderTO } from '../../types/orderTO';
+import { orderTO } from '../../shared/types/orderTO';
 import { BackendCommunicationService } from '../../services/backend-communication.service';
 import { catchError, of } from 'rxjs';
