+++ frontend/shop-floor-assistance-frontend/src/app/shared/component-elements/edit-workflow-dialog/edit-workflow-dialog.component.ts
@@ -5,7 +5,7 @@ import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/materia
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
-import { workflowTO } from '../../../types/workflowTO';
+import { workflowTO } from '../../types/workflowTO';
 import { ButtonComponent } from '../button/button.component';
 
 @Component({
