+++ frontend/shop-floor-assistance-frontend/src/app/shared/component-elements/order-table/order-table.component.ts
@@ -1,7 +1,7 @@
 import { AfterViewInit, Component, EventEmitter, inject, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild, ViewEncapsulation } from '@angular/core';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
-import { orderTO } from '../../../types/orderTO';
+import { orderTO } from '../../types/orderTO';
 import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
 import { MatSort, MatSortModule, Sort } from '@angular/material/sort';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
