+++ frontend/shop-floor-assistance-frontend/src/app/shared/component-elements/workflow-accordion/workflow-accordion.component.ts
@@ -2,12 +2,12 @@ import { CommonModule } from '@angular/common';
 import { AfterViewInit, ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatExpansionModule } from '@angular/material/expansion';
-import { orderTO } from '../../../types/orderTO';
+import { orderTO } from '../../types/orderTO';
 import { MatIconModule } from '@angular/material/icon';
-import { workflowStates } from '../workflowUI-state';
-import { workflowTO } from '../../../types/workflowTO';
+import { workflowStates } from '../../types/workflowUI-state';
+import { workflowTO } from '../../types/workflowTO';
 import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
-import { taskTO } from '../../../types/taskTO';
+import { taskTO } from '../../types/taskTO';
 import { CdkDragDrop, DragDropModule, moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
 import { MatDialog } from '@angular/material/dialog';
 import { EditWorkflowDialogComponent } from '../edit-workflow-dialog/edit-workflow-dialog.component';
