+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/OperatorExecutionTO.ts
