+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/equipmentTO.ts
