+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/itemDropEventType.ts
