+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/itemTO.ts
