+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/login-state.ts
