+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/taskTO.ts
