+++ frontend/shop-floor-assistance-frontend/src/app/shared/types/workflowUI-state.ts
